+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * Moved various palette-related buttons, to group them with the other palette options.
 * A VPK file can be now placed into `vpk_override` to have it be combined with the built VPK. 
 * Added a warning screen to the developer tab of the options window, indicating that some of these options can temporarily break BEE if changed.
+* #1984: The "save settings to palette" will now change to match the selected palette. This means clicking "save" won't accidentally delete saved settings.
 
 ### UCP-Relevant Changes:
 * Terminology change: "Condition Flags" have been renamed to "Condition Tests" - I think that sounds a bit better.